Allow the registration of worker types
A stub method is created for each worker type that is callable from the
native code that returns a new instance of the type, allowing for
dependency injection to be performed. These stub methods are then
registered in a single method call, that can be called by the native
code during startup.